# ast: Remove an unnecessary recursion in Compare with *Term types.