Notable improvements are:

benchmark                                        old ns/op      new ns/op      delta
BenchmarkObjectLookup/5-8                        42.3           37.3           -11.82%
BenchmarkObjectLookup/50-8                       54.1           46.1           -14.79%
BenchmarkObjectLookup/500-8                      55.0           46.0           -16.36%
BenchmarkObjectLookup/5000-8                     51.1           46.5           -9.00%
BenchmarkSetIntersection/5-8                     2340           2280           -2.56%
BenchmarkSetIntersection/50-8                    26239          25365          -3.33%
BenchmarkSetIntersection/500-8                   276254         261412         -5.37%
BenchmarkSetIntersection/5000-8                  3050602        2880013        -5.59%
BenchmarkSetIntersectionDifferentSize/4-8        1881           1801           -4.25%
BenchmarkSetIntersectionDifferentSize/50-8       1978           1870           -5.46%
BenchmarkSetIntersectionDifferentSize/500-8      1969           1862           -5.43%
BenchmarkSetIntersectionDifferentSize/5000-8     1996           1895           -5.06%
BenchmarkSetMembership/5-8                       361            347            -3.88%
BenchmarkSetMembership/50-8                      399            388            -2.76%
BenchmarkSetMembership/500-8                     434            419            -3.46%
BenchmarkSetMembership/5000-8                    460            440            -4.35%
BenchmarkPartialEval/1-8                         9572           9266           -3.20%
BenchmarkPartialEval/10-8                        9634           9310           -3.36%
BenchmarkPartialEval/100-8                       10358          10117          -2.33%
BenchmarkPartialEval/1000-8                      10674          10381          -2.74%
BenchmarkPartialEvalCompile/1-8                  4200311        4037594        -3.87%
BenchmarkPartialEvalCompile/10-8                 6882760        6696489        -2.71%
BenchmarkPartialEvalCompile/100-8                60144772       58103030       -3.39%
BenchmarkPartialEvalCompile/1000-8               3361430791     3325901936     -1.06%
BenchmarkWalk/100-8                              1032722        1030846        -0.18%
BenchmarkWalk/1000-8                             1156007        1106600        -4.27%
BenchmarkWalk/2000-8                             1230263        1205574        -2.01%
BenchmarkWalk/3000-8                             1329760        1289130        -3.06%

Signed-off-by: Teemu Koponen <koponen@styra.com>

<!--

Thanks for submitting a PR to OPA!

Before pressing 'Create pull request' please read the checklist below.

* All code changes should be accompanied with tests. If you are not
modifying any tests, just provide a short explanation of why updates
to tests are not necessary. In addition to helping catch bugs, tests
are extremely helpful in providing _context_ that explains how your
changes can be used.

* All changes to public APIs **must** be accompanied with
docs. Examples of public APIs include built-in functions,
config fields, and of course, exported Go types/functions/constants/etc.

* Commit messages should explain _why_ you made the changes, not what
you changed. Use active voice. Keep the subject line under 50
characters or so.

* All commits must be signed off by the author. If you are not
familiar with signing off, see CONTRIBUTING.md below.

For more information on contributing to OPA see:

* [CONTRIBUTING.md](https://github.com/open-policy-agent/opa/blob/master/CONTRIBUTING.md)
  for high-level contribution guidelines.

* [DEVELOPMENT.md](https://github.com/open-policy-agent/opa/blob/master/docs/devel/DEVELOPMENT.md)
  for development workflow and environment setup.

-->